refactor(web): give the inbox's settings pane its own component
